Add load more changelog pagination
