+++ apps/www/pages/changelog.tsx
@@ -1,127 +1,14 @@
-import { ArrowLeftIcon, ArrowRightIcon } from '@heroicons/react/outline'
-import { createAppAuth } from '@octokit/auth-app'
-import { Octokit } from '@octokit/core'
-import { paginateGraphql } from '@octokit/plugin-paginate-graphql'
-import { Octokit as OctokitRest } from '@octokit/rest'
 import dayjs from 'dayjs'
 import { GitCommit } from 'lucide-react'
 import { GetServerSideProps } from 'next'
-import { MDXRemote, MDXRemoteSerializeResult } from 'next-mdx-remote'
+import { MDXRemote } from 'next-mdx-remote'
 import { NextSeo } from 'next-seo'
 import Link from 'next/link'
+import { useState } from 'react'
 import CTABanner from '~/components/CTABanner'
 import DefaultLayout from '~/components/Layouts/Default'
-import { deletedDiscussions } from '~/lib/changelog.utils'
+import { fetchChangelogPage, recursiveDecodeURI, type Entry } from '~/lib/changelog.server'
 import mdxComponents from '~/lib/mdx/mdxComponents'
-import { mdxSerialize } from '~/lib/mdx/mdxSerialize'
-
-export type Discussion = {
-  id: string
-  updatedAt: string
-  url: string
-  title: string
-  body: string
-}
-
-type Entry = {
-  id: string
-  title: string
-  url: string
-  created_at: string
-  source: MDXRemoteSerializeResult
-  type: string
-}
-
-export type DiscussionsResponse = {
-  repository: {
-    discussions: {
-      totalCount: number
-      nodes: Discussion[]
-      pageInfo: any
-    }
-  }
-}
-
-// uses the graphql api
-async function fetchDiscussions(
-  owner: string,
-  repo: string,
-  categoryId: string,
-  cursor: string | null = null
-) {
-  const ExtendedOctokit = Octokit.plugin(paginateGraphql)
-  type ExtendedOctokit = InstanceType<typeof ExtendedOctokit>
-
-  const octokit = new ExtendedOctokit({
-    authStrategy: createAppAuth,
-    auth: {
-      appId: process.env.GITHUB_CHANGELOG_APP_ID,
-      installationId: process.env.GITHUB_CHANGELOG_APP_INSTALLATION_ID,
-      privateKey: process.env.GITHUB_CHANGELOG_APP_PRIVATE_KEY,
-    },
-  })
-
-  const query = `
-    query troubleshootDiscussions($cursor: String, $owner: String!, $repo: String!, $categoryId: ID!) {
-      repository(owner: $owner, name: $repo) {
-        discussions(first: 50, after: $cursor, categoryId: $categoryId, orderBy: { field: CREATED_AT, direction: DESC }) {
-          totalCount
-          pageInfo {
-            hasPreviousPage
-            hasNextPage
-            startCursor
-            endCursor
-          }
-          nodes {
-            id
-            publishedAt
-            createdAt
-            url
-            title
-            body
-          }
-        }
-      }
-    }
-  `
-  const queryVars = {
-    owner,
-    repo,
-    categoryId,
-    cursor,
-  }
-
-  // fetch discussions
-  const {
-    repository: {
-      discussions: { nodes: discussions, pageInfo },
-    },
-  } = await octokit.graphql<DiscussionsResponse>(query, queryVars)
-
-  return { discussions, pageInfo }
-}
-
-function isEncoded(uri: string | null | undefined) {
-  uri = uri ?? ''
-  return uri !== decodeURIComponent(uri)
-}
-
-// Decodes a URI if it is encoded
-const recursiveDecodeURI = (uri: string | null) => {
-  if (!uri) {
-    return uri
-  }
-  let tries = 0
-  while (isEncoded(uri)) {
-    uri = decodeURIComponent(uri)
-    tries++
-    if (tries > 10) {
-      break
-    }
-  }
-
-  return uri
-}
 
 /**
  * [Terry]
@@ -140,122 +27,13 @@ export const getServerSideProps: GetServerSideProps = async ({ res, query }) => 
   // in some cases the next cursor is encoded twice or more times due to the user pasting the url, so we need to decode it multiple times.
   const next = recursiveDecodeURI(encodedNext)
   const restPage = query.restPage ? Number(query.restPage) : 1
-
-  const octokitRest = new OctokitRest({
-    auth: process.env.GITHUB_CHANGELOG_APP_REST_KEY,
-  })
-
-  // uses the rest api
-  async function fetchGitHubReleases() {
-    try {
-      const response = await octokitRest.repos.listReleases({
-        owner: 'supabase',
-        repo: 'supabase',
-        per_page: 10,
-        page: restPage,
-      })
-
-      return response.data || []
-    } catch (error) {
-      console.error(error)
-      return []
-    }
-  }
-
-  // Process as of Feb. 2024:
-  // create a Release each month and create a corresponding changelog discussion
-  // — we don't want to pull in both the changelog entry and the release entry
-  // — we want to ignore new releases and only show the old ones that don't have a corresponding changelog discussion
-  // — so we have this list of old releases that we want to show
-  const oldReleases = [
-    40981345, 39091930, 37212777, 35927141, 34612423, 33383788, 32302703, 30830915, 29357247,
-    28108378,
-  ]
-
-  const releases = (await fetchGitHubReleases()).filter(
-    (release) => release.id && oldReleases.includes(release.id)
-  )
-
-  const { discussions, pageInfo } = await fetchDiscussions(
-    'supabase',
-    'supabase',
-    'DIC_kwDODMpXOc4CAFUr', // 'Changelog' category
-    next
-  )
-
-  if (!discussions) {
-    return {
-      props: {
-        notFound: true,
-      },
-    }
-  }
-
-  // Process discussions
-  const formattedDiscussions = await Promise.all(
-    discussions.map(async (item: any): Promise<any> => {
-      try {
-        const discussionsMdxSource: MDXRemoteSerializeResult = await mdxSerialize(item.body)
-        // Find a date rewrite for the current item's title
-        const dateRewrite = deletedDiscussions.find((rewrite) => {
-          return item.title && rewrite.title && item.title.includes(rewrite.title)
-        })
-
-        // Use the createdAt date from dateRewrite if found, otherwise use item.createdAt
-        const created_at = dateRewrite ? dateRewrite.createdAt : item.createdAt
-
-        return {
-          ...item,
-          source: discussionsMdxSource,
-          type: 'discussion',
-          created_at,
-          url: item.url,
-        }
-      } catch (err) {
-        console.error(`Problem processing discussion MDX: ${err}`)
-      }
-    })
-  )
-
-  // Process releases
-  const formattedReleases = await Promise.all(
-    releases.map(async (item: any): Promise<any> => {
-      try {
-        const releasesMdxSource: MDXRemoteSerializeResult = await mdxSerialize(item.body)
-
-        return {
-          ...item,
-          source: releasesMdxSource,
-          type: 'release',
-          created_at: item.created_at,
-          title: item.name ?? '',
-          url: item.html_url ?? '',
-        }
-      } catch (err) {
-        console.error(`Problem processing discussion MDX: ${err}`)
-      }
-    })
-  )
-
-  // Combine discussions and releases into a single array of entries
-  const combinedEntries = formattedDiscussions.concat(formattedReleases).filter(Boolean)
-
-  const sortedCombinedEntries = combinedEntries.sort((a: any, b: any) => {
-    const dateA = dayjs(a.created_at)
-    const dateB = dayjs(b.created_at)
-
-    if (dateA.isValid() && dateB.isValid()) {
-      return dateB.diff(dateA)
-    } else {
-      return 0
-    }
-  })
+  const data = await fetchChangelogPage({ next, restPage })
 
   return {
     props: {
-      changelog: sortedCombinedEntries,
-      pageInfo: pageInfo,
-      restPage: Number(restPage),
+      changelog: data.changelog,
+      pageInfo: data.pageInfo,
+      restPage: data.restPage,
     },
   }
 }
@@ -267,7 +45,40 @@ interface ChangelogPageProps {
 }
 
 function ChangelogPage({ changelog, pageInfo, restPage }: ChangelogPageProps) {
-  const { endCursor: end, hasNextPage, hasPreviousPage } = pageInfo
+  const [entries, setEntries] = useState<Entry[]>(changelog)
+  const [currentPageInfo, setCurrentPageInfo] = useState(pageInfo)
+  const [currentRestPage, setCurrentRestPage] = useState(restPage)
+  const [isLoadingMore, setIsLoadingMore] = useState(false)
+  const { endCursor: end, hasNextPage } = currentPageInfo ?? {}
+
+  const handleLoadMore = async () => {
+    if (!hasNextPage || isLoadingMore) {
+      return
+    }
+
+    setIsLoadingMore(true)
+    try {
+      const params = new URLSearchParams()
+      if (end) {
+        params.set('next', end)
+      }
+      params.set('restPage', String(currentRestPage + 1))
+
+      const response = await fetch(`/api/changelog?${params.toString()}`)
+      if (!response.ok) {
+        throw new Error('Failed to load more changelog entries')
+      }
+
+      const data = await response.json()
+      setEntries((prev) => prev.concat(data.changelog ?? []))
+      setCurrentPageInfo(data.pageInfo ?? currentPageInfo)
+      setCurrentRestPage(data.restPage ?? currentRestPage + 1)
+    } catch (error) {
+      console.error(error)
+    } finally {
+      setIsLoadingMore(false)
+    }
+  }
 
   const TITLE = 'Changelog'
   const DESCRIPTION = 'New updates and improvements to Supabase'
@@ -298,8 +109,8 @@ function ChangelogPage({ changelog, pageInfo, restPage }: ChangelogPageProps) {
 
           {/* Content */}
           <div className="grid gap-12 lg:gap-36">
-            {changelog.length > 0 &&
-              changelog
+            {entries.length > 0 &&
+              entries
                 .filter((entry: Entry) => !entry.title.includes('[d]'))
                 .map((entry: Entry, i: number) => {
                   return (
@@ -333,21 +144,18 @@ function ChangelogPage({ changelog, pageInfo, restPage }: ChangelogPageProps) {
                   )
                 })}
           </div>
-          <div className="my-8 flex items-center gap-4">
-            {hasPreviousPage && (
-              <Link href={`/changelog`} className="flex items-center gap-2">
-                <ArrowLeftIcon width={14} /> Previous
-              </Link>
-            )}
-            {hasNextPage && (
-              <Link
-                href={`/changelog?next=${end}&restPage=${restPage + 1}`}
-                className="flex items-center gap-2"
+          {hasNextPage && (
+            <div className="my-8 flex items-center justify-center">
+              <button
+                type="button"
+                onClick={handleLoadMore}
+                disabled={isLoadingMore}
+                className="border-control text-foreground bg-foreground/5 hover:bg-foreground/10 flex items-center gap-2 rounded-md border px-4 py-2 text-sm transition disabled:cursor-not-allowed disabled:opacity-60"
               >
-                Next <ArrowRightIcon width={14} />
-              </Link>
-            )}
-          </div>
+                {isLoadingMore ? 'Loading...' : 'Load more'}
+              </button>
+            </div>
+          )}
         </div>
 
         <CTABanner />
